requirements.txt
PyYAML==6.0.2
PyYAML==6.0.2
spotipy==2.24.0
yt_dlp==2024.8.6
